summarize-emails/requirements.txt
openai>=1.0.0
markdown2
google-api-python-client
google-auth
google-auth-oauthlib
